Rename api.py to streams.py
