+++ airbyte-integrations/connectors/source-sendgrid/source_sendgrid/source.py
@@ -29,7 +29,7 @@ from airbyte_cdk.sources import AbstractSource
 from airbyte_cdk.sources.streams import Stream
 from airbyte_cdk.sources.streams.http.auth import TokenAuthenticator
 
-from .api import (
+from .streams import (
     Blocks,
     Bounces,
     Campaigns,
